Create CreateSuiteRequest (#845)
