+++ src/Request/SuiteRequest.php
@@ -10,12 +10,10 @@ class SuiteRequest
     public const PARAMETER_TESTS = 'tests';
 
     /**
-     * @param non-empty-string             $sourceId
      * @param non-empty-string             $label
      * @param array<int, non-empty-string> $tests
      */
     public function __construct(
-        public readonly string $sourceId,
         public readonly string $label,
         public readonly array $tests,
     ) {
